requirements.txt
streamlit
pandas
numpy
plotly
supabase
google-generativeai
python-dotenv
fpdf2
